Moved `encode` to new utils directory
