+++ lib/prismic-client/src/utils/encode.spec.ts
