+++ lib/prismic-client/src/utils/encode.ts
